Adpatation of configuration files to the new architecture

+++ veinsBDI/src/base/manager/Trigger/Trigger.h
@@ -35,7 +35,7 @@ private:
     typedef struct _trigger_sequence{
         struct _trigger_sequence *next;
         DataType type;
-        uint8_t dataLength;
+        uint32_t dataLength;
         void *dataSequence;
     } TriggerSequence;
